backend/requirements.txt
# Django Core
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0

# Database
psycopg2-binary>=2.9
dj-database-url>=2.0

# Authentication
djangorestframework-simplejwt>=5.3
django-filter>=23.0

# Supabase Storage
supabase>=2.0

# Audio Processing
librosa==0.10.2.post1
numpy==1.26.4
scipy==1.11.4
pydub==0.25.1
soundfile>=0.12

# NLP & AI
g2p_en==2.1.0
transformers==4.38.2
torch==2.2.2
torchaudio==2.2.2

# LLM Integration
groq>=0.4
google-generativeai>=0.5
cerebras-cloud-sdk>=1.0
